+++ client/src/layout/MainLayout/Header/HeaderContent/jobs.jsx
@@ -95,7 +95,7 @@ const Jobs = () => {
         }, 5000);
 
         return () => clearInterval(interval);
-    }, []);
+    }, [open, isAdmin]);
 
     
     const handleToggle = () => {