+++ changelog.md
@@ -4,6 +4,7 @@
  - Re-Implements the SSO using openID internally - fixes issue where you need to re-loging when app are on different domains (because of browser cookies limitations)
  - Implements local HTTPS Certificate Authority, to locally trust self-signed certificates on devices
  - Added new folder button to file picker
+ - Cosmos now waits for CRON jobs to be over before restarting the server
  - Fixed bug with RClone storage duplication in the UI
  - Implements hybrid HTTPS with public and self-signed certificates switched on the fly
  - OpenID now returns more info in case of errors when Cosmos is in debug mode
@@ -11,7 +12,8 @@
  - Improved local IP detection (Thanks @r41d)
  - Updated LEGO to 4.21.0
  - Fixed file picker prefix issue in docker container
- - Fix RClone not starting
+ - Added OpenID IDTokenSigningAlgValuesSupported
+ - Fix RClone not starting (hopefully)
 
 ## Version 0.17.7
  - Fix error code on login screen